Add pl006
