--- cmd/cl006/ac000/zz_generated.command.go
@@ -1,86 +1,0 @@
-package ac000
-
-import (
-	"context"
-	"io"
-	"os"
-	"strings"
-
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"github.com/spf13/cobra"
-
-	"github.com/giantswarm/awscnfm/v12/pkg/action"
-	"github.com/giantswarm/awscnfm/v12/pkg/action/cl006/ac000"
-	"github.com/giantswarm/awscnfm/v12/pkg/config"
-	"github.com/giantswarm/awscnfm/v12/pkg/env"
-)
-
-const (
-	name        = "ac000"
-	description = "Execute action ac000 for cluster cl006."
-)
-
-type Config struct {
-	Logger micrologger.Logger
-	Stderr io.Writer
-	Stdout io.Writer
-}
-
-func New(config Config) (*cobra.Command, error) {
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.Stderr == nil {
-		config.Stderr = os.Stderr
-	}
-	if config.Stdout == nil {
-		config.Stdout = os.Stdout
-	}
-
-	f := &flag{}
-
-	r := &runner{
-		flag:   f,
-		logger: config.Logger,
-		stderr: config.Stderr,
-		stdout: config.Stdout,
-	}
-
-	c := &cobra.Command{
-		Use:   name,
-		Short: description,
-		Long:  mustLong(),
-		RunE:  r.Run,
-	}
-
-	f.Init(c)
-
-	return c, nil
-}
-
-func mustLong() string {
-	ctx := context.Background()
-
-	var err error
-
-	var e action.Explainer
-	{
-		c := ac000.ExplainerConfig{
-			Scope:         "cl006",
-			TenantCluster: config.Cluster("cl006", env.TenantCluster()),
-		}
-
-		e, err = ac000.NewExplainer(c)
-		if err != nil {
-			panic(err)
-		}
-	}
-
-	s, err := e.Explain(ctx)
-	if err != nil {
-		panic(err)
-	}
-
-	return strings.TrimSpace(s)
-}
