+++ cmd/plan/command.go
@@ -6,6 +6,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/giantswarm/awscnfm/v12/cmd/plan/pl001"
+	"github.com/giantswarm/awscnfm/v12/cmd/plan/pl006"
 )
 
 const (
@@ -36,6 +37,18 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
+	var pl006Cmd *cobra.Command
+	{
+		c := pl006.Config{
+			Logger: config.Logger,
+		}
+
+		pl006Cmd, err = pl006.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
 	f := &flag{}
 
 	r := &runner{
@@ -53,6 +66,7 @@ func New(config Config) (*cobra.Command, error) {
 	f.Init(c)
 
 	c.AddCommand(pl001Cmd)
+	c.AddCommand(pl006Cmd)
 
 	return c, nil
 }
