+++ cmd/plan/pl006/error.go
@@ -1,4 +1,4 @@
-package ac000
+package pl006
 
 import "github.com/giantswarm/microerror"
 
@@ -11,11 +11,11 @@ func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
 }
 
-var invalidFlagsError = &microerror.Error{
-	Kind: "invalidFlagsError",
+var invalidFlagError = &microerror.Error{
+	Kind: "invalidFlagError",
 }
 
-// IsInvalidFlags asserts invalidFlagsError.
-func IsInvalidFlags(err error) bool {
-	return microerror.Cause(err) == invalidFlagsError
+// IsInvalidFlag asserts invalidFlagsError.
+func IsInvalidFlag(err error) bool {
+	return microerror.Cause(err) == invalidFlagError
 }
