+++ cmd/plan/pl006/flag.go
@@ -1,6 +1,8 @@
-package ac000
+package pl006
 
-import "github.com/spf13/cobra"
+import (
+	"github.com/spf13/cobra"
+)
 
 type flag struct {
 }
