+++ cmd/plan/pl006/plan.go
@@ -1,4 +1,4 @@
-package ac000
+package pl006
 
 import (
 	"time"
@@ -7,36 +7,32 @@ import (
 )
 
 // Plan describes in which order and with which tolerance to execute actions of
-// this cluster scope.
+// this test plan.
 var Plan = []plan.Step{
 	{
-		Action:  "ac001",
+		Action:  "create/cluster/singlecontrolplane",
 		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
-		Comment: "create cluster CRs",
 	},
 	{
-		Action:  "ac002",
+		Action:  "verify/cluster/created",
 		Backoff: plan.NewBackoff(30*time.Minute, 3*time.Minute),
-		Comment: "check cluster access",
 	},
 	{
-		Action:  "ac003",
+		Action:  "update/cluster/ha",
 		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
 		Comment: "trigger upgrade to HA",
 	},
 	{
-		Action:  "ac004",
-		Backoff: plan.NewBackoff(120*time.Minute, 10*time.Minute),
-		Comment: "check upgrade successful",
-	},
-	{
-		Action:  "ac005",
+		Action:  "verify/cluster/hasetup",
 		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
-		Comment: "delete cluster CRs",
+		Comment: "trigger upgrade to HA",
 	},
 	{
-		Action:  "ac006",
+		Action:  "delete/cluster",
+		Backoff: plan.NewBackoff(10*time.Second, 2*time.Second),
+	},
+	{
+		Action:  "verify/cluster/deleted",
 		Backoff: plan.NewBackoff(90*time.Minute, 9*time.Minute),
-		Comment: "check CRs deleted",
 	},
 }
